requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
SQLAlchemy>=2.0
pydantic-settings>=2.4
python-dotenv>=1.0
requests>=2.31
beautifulsoup4>=4.12
pytest>=8.0

